Minor fixes & refactoring
Libs:
* updated mbedtls to 2.4.0-apache

Common:
* DataCallback now has correct signature void(data::Value &&)

Serenity:
* Fixed output filter looping in http/0.9

Stappler:
* _token prefix for localizable tokens
* Log: fixed vsnprintf usage on linux
* MemLeaks: new util for memory leaks detection
* Rc/Arc: fixed memleaks
* Asset: flags now started from 2 to prevent mixes with Initial value
* FontFormatter: added ranged reconstruction function
* rich_text::Document: removed unused functions
* rich_text::Source: fixed Document double-loading
* DynamicLabel: background quads rendering disabled

Material:
* TextField: fixed cursor position
* FloatingMenu: menu strings now parsed as localizable
* RecyclerScroll: first item animation fixed
* RecyclerScroll: removal now in reversed order to prevent corruption

+++ stappler/src/features/rich_text/renderer/SPRichTextLayout.h
@@ -93,11 +93,11 @@ struct Label {
 	font::FormatSpec format;
 	float height = 0.0f;
 
-	Rect getLineRect(uint16_t lineId, float density, const Vec2 & = Vec2()) const;
+	Rect getLineRect(uint32_t lineId, float density, const Vec2 & = Vec2()) const;
 	Rect getLineRect(const font::LineSpec &, float density, const Vec2 & = Vec2()) const;
 
-	uint16_t getLineForCharId(uint16_t id) const;
-	Vector<Rect> getLabelRects(uint16_t first, uint16_t last, float density, const Vec2 & = Vec2(), const Padding &p = Padding()) const;
+	uint32_t getLineForCharId(uint32_t id) const;
+	Vector<Rect> getLabelRects(uint32_t first, uint32_t last, float density, const Vec2 & = Vec2(), const Padding &p = Padding()) const;
 };
 
 struct Object {
